requirements-core.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
h11==0.14.0
idna==3.8
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
